+++ CHANGELOG.md
@@ -112,6 +112,8 @@
   [#9877](https://github.com/Kong/kong/pull/9877)
 - **JWT**: Deny requests that have different tokens in the jwt token search locations. Thanks Jackson 'Che-Chun' Kuo from Latacora for reporting this issue.
   [#9946](https://github.com/Kong/kong/pull/9946)
+- **Statsd**: Fix a bug that the batch queue in statsd can't get the right result when processing batch entries produced by this plugin.
+  [#10052](https://github.com/Kong/kong/pull/10052)
 
 ### Dependencies